% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{as_graph}
\alias{as_graph}
\title{Get igraph object from a dataframe representing package dependencies}
\usage{
as_graph(dependency_df)
}
\arguments{
\item{dependency_df}{(\code{dataframe}) Representing package dependencies}
}
\value{
igraph object
}
\description{
Packages dependencies (typically obtained from
\code{\link[=get_dependencies]{get_dependencies()}} or \code{\link[=get_neighborhood]{get_neighborhood()}}) is converted into a
\href{https://cran.r-project.org/package=igraph}{igraph} object.
}
\examples{
\dontrun{
init()
as_graph(get_neighborhood("mlr3"))
}
}
