% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{deptable}
\alias{deptable}
\title{deptable}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 55699 rows and 3 columns.}
\usage{
deptable
}
\description{
(tibble) A tibble with three columns: 'pkg_1', 'relation' and
  'pkg_2'. Every row defines a dependency. This is computed for all packages
  in 'packmeta'
}
\keyword{datasets}
