% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direct.R
\name{get_imports}
\alias{get_imports}
\title{get_imports}
\usage{
get_imports(packages, level = 1L)
}
\arguments{
\item{packages}{(non-empty character vector) Package names}

\item{level}{(positive integer) Depth of recursive dependency}
}
\value{
A tibble with three columns: `pkg_1`, `relation` and `pkg_2`
}
\description{
Get dependencies
}
\examples{
pkggraph::init()
pkggraph::get_imports("dplyr")
}
\seealso{
\code{\link{get_depends}}, \code{\link{get_imports}},
  \code{\link{get_linkingto}}, \code{\link{get_suggests}},
  \code{\link{get_enhances}}, \code{\link{get_all_dependencies}},
  \code{\link{get_reverse_imports}}
}
\author{
Srikanth KS
}
