% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indirect.R
\name{get_reverse_suggests}
\alias{get_reverse_suggests}
\title{get_reverse_suggests}
\usage{
get_reverse_suggests(packages, level = 1L)
}
\arguments{
\item{packages}{(non-empty character vector) Package names}

\item{level}{(positive integer) Depth of recursive dependency}
}
\value{
A tibble with three columns: `pkg_1`, `relation` and `pkg_2`
}
\description{
Get reverse dependencies
}
\examples{
pkggraph::init()
pkggraph::get_reverse_suggests("purrr")
}
\seealso{
\code{\link{get_reverse_depends}}, \code{\link{get_reverse_imports}},
  \code{\link{get_reverse_linkingto}}, \code{\link{get_reverse_suggests}},
  \code{\link{get_reverse_enhances}}, \code{\link{get_all_reverse_dependencies}},
  \code{\link{get_suggests}}
}
\author{
Srikanth KS
}
