% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infix.R
\name{\%enhances\%}
\alias{\%enhances\%}
\title{Check enhances}
\usage{
pkg_1 \%enhances\% pkg_2
}
\arguments{
\item{pkg_1}{a package name}

\item{pkg_2}{a package name}
}
\value{
TRUE or FALSE
}
\description{
Check whether pkg_1 has a dependency on pkg_2
}
\examples{
pkggraph::init(local = TRUE)
"dplyr" \%enhances\% "tibble"
}
\author{
Srikanth KS
}
