% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infix.R
\name{\%imports\%}
\alias{\%imports\%}
\title{Check imports}
\usage{
pkg_1 \%imports\% pkg_2
}
\arguments{
\item{pkg_1}{a package name}

\item{pkg_2}{a package name}
}
\value{
TRUE or FALSE
}
\description{
Check whether pkg_1 has a dependency on pkg_2
}
\examples{
pkggraph::init(local = TRUE)
"dplyr" \%imports\% "tibble"
}
\author{
Srikanth KS
}
