% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infix.R
\name{\%linkingto\%}
\alias{\%linkingto\%}
\title{Check linkingto}
\usage{
pkg_1 \%linkingto\% pkg_2
}
\arguments{
\item{pkg_1}{a package name}

\item{pkg_2}{a package name}
}
\value{
TRUE or FALSE
}
\description{
Check whether pkg_1 has a dependency on pkg_2
}
\examples{
pkggraph::init()
"dplyr" \%linkingto\% "tibble"
}
\author{
Srikanth KS
}
