% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infix.R
\name{\%relies\%}
\alias{\%relies\%}
\title{Check relies}
\usage{
pkg_1 \%relies\% pkg_2
}
\arguments{
\item{pkg_1}{(string) A package name}

\item{pkg_2}{(string) A package name}
}
\value{
(flag) TRUE, if 'pkg_1' `relies` on 'pkg_2'
}
\description{
Check whether a package has a recursive dependency on the other
}
\examples{
pkggraph::init()
"dplyr" \%relies\% "tibble"
}
\seealso{
\code{\link{relies}}, \code{\link{reverse_relies}}
}
\author{
Srikanth KS
}
