% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{packmeta}
\alias{packmeta}
\title{packmeta}
\format{An object of class \code{matrix} with 10718 rows and 17 columns.}
\usage{
packmeta
}
\description{
(A character matrix) Output of 'utils::available.packages'
}
\keyword{datasets}
