% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkggraph-package.R
\docType{package}
\name{pkggraph-package}
\alias{pkggraph}
\alias{pkggraph-package}
\title{pkggraph: Explore the R Package Dependencies on the Comprehensive R Archive Network (CRAN) Like Repositories}
\description{
Explore various dependencies of a packages (on the Comprehensive R Archive Network Like repositories). The functions get_neighborhood() and get_dependencies() provide dependencies of packages and as_graph() can be used to convert into a 'igraph' object for further analysis and plotting.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/talegari/pkggraph}
  \item Report bugs at \url{https://github.com/talegari/pkggraph/issues}
}

}
\author{
\strong{Maintainer}: Komala Sheshachala Srikanth \email{sri.teach@gmail.com}

Authors:
\itemize{
  \item Singh Nikhil \email{nikhilsingh2009@gmail.com}
}

}
\keyword{internal}
