% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{plot.neighborhood_graph}
\alias{plot.neighborhood_graph}
\title{Static plot of package dependencies}
\usage{
\method{plot}{neighborhood_graph}(x, layout = "sugiyama", ...)
}
\arguments{
\item{x}{neighborhood_graph}

\item{layout}{layout is passed to \link[ggraph:ggraph]{ggraph::ggraph}}

\item{...}{Not used}
}
\value{
ggraph plot
}
\description{
Plot of packages dependencies or neighborhood with edges colored
by \code{relation} and node sized by 'centrality' from a \code{neighborhood_graph}
obtained from \link{as_graph}.
}
\examples{
\dontrun{
init()
get_neighborhood("mboost") |>
  as_graph() |>
  plot()
}
}
