% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rely.R
\name{relies}
\alias{relies}
\title{relies}
\usage{
relies(packages)
}
\arguments{
\item{packages}{(non-empty character vector) Package names}
}
\value{
(Named list) A name is the package name from 'packages'. A Value is a
  character vector of all packages which the package 'relies' (Captures
  recursive dependencies of these types: "Depends", "Imports", "LinkingTo")
}
\description{
Captures recursive dependencies of these types: "Depends",
  "Imports", "LinkingTo"
}
\examples{
pkggraph::init(local = TRUE)
pkggraph::relies("tidytext")
}
\seealso{
\code{\link{reverse_relies}}
}
\author{
Srikanth KS
}
