% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rely.R
\name{reverse_relies}
\alias{reverse_relies}
\title{reverse_relies}
\usage{
reverse_relies(packages)
}
\arguments{
\item{packages}{(non-empty character vector) Package names}
}
\value{
(Named list) A name is the package name from 'packages'. A Value is a
  character vector of all packages which the package 'relies' (Captures
  reverse recursive dependencies of these types: "Depends", "Imports",
  "LinkingTo")
}
\description{
Captures reverse recursive dependencies of these types:
  "Depends", "Imports", "LinkingTo"
}
\examples{
pkggraph::init(local = TRUE)
pkggraph::reverse_relies("data.table")
}
\seealso{
\code{\link{relies}}
}
\author{
Srikanth KS
}
