% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init}
\alias{init}
\title{init}
\usage{
init(local = TRUE, repository = "CRAN", ...)
}
\arguments{
\item{local}{(flag, default: TRUE) If \itemize{ \item FALSE: Tries to to
download package data from CRAN over internet and compute dependencies
\item TRUE: Loads data that comes with the package corresponding to 1st
June 2017 17:14 IST}}

\item{repository}{(character vector, Default: "CRAN") One among c("CRAN",
"BioCsoft", "BioCann", "BioCexp", "BioCextra", "omegahat"). To use a
repository not in this list, set 'repository' to NULL and pass named
argument called 'repos' with a valid repository address. This will be
passed as is to `utils::available.packages()`.}

\item{...}{Additional parameters to be passed to `available.packages()`}
}
\value{
An invisible TRUE
}
\description{
Initiate the package by loading the data into parent frame. This
  should be done as soon as the package is loaded or attached. This
  creates(rewrites) new variables 'deptable' and 'packmeta' to the
  environment where it is run from.
}
\author{
Srikanth KS
}
