% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkggraph_package.R
\docType{package}
\name{pkggraph-package}
\alias{pkggraph}
\alias{pkggraph-package}
\title{pkggraph}
\description{
Interactively explore various dependencies of a package(s) (on the Comprehensive R Archive Network Like repositories) and perform analysis using tidy philosophy. Most of the functions return a 'tibble' object (enhancement of 'dataframe') which can be used for further analysis. The package offers functions to produce 'network' and 'igraph' dependency graphs. The 'plot' method produces a static plot based on 'ggnetwork' and 'plotd3' function produces an interactive D3 plot based on 'networkD3'.
}
\details{
See the vignette for further details
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/talegari/pkggraph}
  \item Report bugs at \url{https://github.com/talegari/pkggraph/issues}
}

}
\author{
\strong{Maintainer}: KS Srikanth \email{sri.teach@gmail.com}

Authors:
\itemize{
  \item Singh Nikhil \email{nikhilsingh2009@gmail.com}
}

}
